print array

+++ 0x05-pointers_arrays_strings/8-print_array.c
@@ -12,7 +12,7 @@ void print_array(int *a, int n)
 {
 	int len, i;
 
-	len = sizeof(a) / sizeof(a[0]);
+	len = (int)(sizeof(a) / sizeof(a[0]));
 	for (i = 0; i < n && i < len ; i++)  
 	{
 		if (i == n - 1 || i == len -1)
